minor cleanup, diagram orientation, rm svg/
The power supply diagram has been re-oriented
so it will be right side up when the landscape has
the right side of the page.  Previously it was
the opposite.

Moved a power supply line in the control diagram
outside of its control group so it was clear that
it comes from somewhere else.

Removed svg/ which contained the out of date
Inkscape diagrams.