created schematic in Kicad
The schematic has been converted to Kicad.
No errors were found by the Electronic Rules
Checker (ERC).

Several features of Kicad helped make the schematic
simple.  It has the concept of "sheets" which allow
a sub circuit to be created and interfaced through
its inputs/outputs.  And this sub circuit can be used
repeatedly.  This made the driver circuit simple
since it was just eight copies of the same circuit.

Next the netlist needs to be created and then
the pcb can be designed.