minor power circuit changes
The MC34063A circuit has been tested using
actual components.  All the calculations were
correct and it behaves expected.  There is
no measurable heat loss which is a dramatic
change from the LM7805 which was hot enough
to cause burns.

The output filter capacitor has been changed
from 182 uF to 180 uF because a 182 is not available.
And a 0.5M bleed resistor has been added in
parallel with the rectifier filter capacitor.