increased power CT capacitor size
It was determined that when the DC source
voltage is low (13.8 volts instead of 17 volts)
the MC34063 will regulate the output voltage
using a higher frequency.  And since a relatively
large inductor (L1, 220 uH) is being used this
can create a substantial voltage drop (1.5 V).

Z_L = 2*pi*f*L

To resolve this problem the timing capacitor (CT)
was doubled in size from 450 pF to 1000 pF.