fixed wasteful process dispatch
The spkpi-dispatch daemon will spawn and child
process anytime the mode file is changed.  The
previous implementation would write to all the
files, include the mode file, even if the value
was the same.  This was very wasteful.

To fix this problem two objects are used: one
from the files and the other from the processed
$_POST variable.  Then only the differences between
these objects are written to the files.