revised Power On Test and Manual Control Test
The Power On Test and Manual Control Test procedures
have been updated.  Some images still need to be added.
Remaining test procedures are in a messy state and need to
be updated.

The previous test procedure assumed the command line was used
but it is now focused on the web interface.